Merged in bugfix/CONNECT-133-php-8-kompatibilität (pull request #1)
Bugfix/CONNECT-133 php 8 kompatibilität

* Set php-platform to 7.4 for composer as required by shopware/core

Added the php 7.4.0 as the platform in composer to comply to the used platform version in shopware/core. Required to run composer install without errors.
* Trying to resolve pipeline build error

Changing localhost to 127.0.0.1 (by recommendation of https://support.atlassian.com/bitbucket-cloud/docs/databases-and-service-containers/) and replacing pipleline database name
* Setting root as default password again

Root was the associated password for the default database before, returning to that after successfull pipeline test
* Reverting changes on .env

* Reverting .env url back to localhost

Not relevant for the pipeline

Approved-by: Joshua Behrens